Rename communication namespace to 'serial'
